Moved OpusUtil to sx.blah.discord.internal
